% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_checks.R
\name{check_rstudio}
\alias{check_rstudio}
\title{Check for correct version of RStudio}
\usage{
check_rstudio(version1 = "1.4.1717", version2021 = "2019.9.0.351")
}
\arguments{
\item{version}{required version number (or higher)}
}
\value{
message with the result
}
\description{
Check for correct version of RStudio
}
\details{
Checks whether the user is using the correct version of RStudio (or higher)
}
\keyword{internal}
