% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{elegans}
\alias{elegans}
\title{elegans network data}
\format{
A data frame with 297 vertices and 2344 edges. Directed, Weighted.
}
\source{
\url{http://networkrepository.com/celegansneural.php}
}
\usage{
data(elegans)
}
\description{
elegans network
}
\details{
elegans network, in \code{igraph}-format. This network is used in the course
text book.
}
\references{
\insertRef{Elegans}{SNA4DS}
}
\keyword{datasets}
