% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{enron}
\alias{enron}
\title{enron network data}
\format{
An object of class \code{igraph} of length 10.
}
\source{
\url{http://konect.cc/networks/enron/} and \url{http://www.cs.cmu.edu/~enron/}
}
\usage{
enron
}
\description{
enron network
}
\details{
enron network, in \code{igraph}-format. This network is used in the course
text book.

The network consists of 321917 emails sent between employees of Enron between
1999 and 2003. Nodes in the network are individual employees and edges are
individual emails.  Edges are directed and weighted. The direction is from
the sender to the receiver and the weight represents the number of messages.
It is possible to send an email to oneself, and thus this network contains
loops.

The edges have a weight attribute.

Note that there are many versions of the enron dataset available on the
internet, this specific dataset is the dataset as used in the text book.
}
\references{
\insertRef{KONECT}{SNA4DS}

\insertRef{ENRONCORPUS}{SNA4DS}
}
\keyword{datasets}
