% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{enwiki}
\alias{enwiki}
\title{enwiki network data}
\format{
A data frame with 15220 vertices with 194103 edges. Directed, Unweighted.
}
\usage{
enwiki
}
\description{
enwiki network
}
\details{
enwiki network, in \code{igraph}-format. This network is used in the course
text book.
}
\keyword{datasets}
