% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{everglades}
\alias{everglades}
\title{everglades network data}
\format{
A data frame with 15220 vertices with 194103 edges. Directed, Unweighted.
}
\source{
\url{http://networkrepository.com/eco-everglades.php}
}
\usage{
data(everglades)
}
\description{
everglades network
}
\details{
everglades network, in \code{igraph}-format. This network is used in the course
text book.

eco-florida also exists and is slightly larger
}
\references{
\insertRef{ulanowicz1998network}{SNA4DS}

\insertRef{melian2004food}{SNA4DS}
}
\keyword{datasets}
