% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GOF.plot.R
\name{GOF.plot}
\alias{GOF.plot}
\title{GOF.plot}
\usage{
GOF.plot(m)
}
\arguments{
\item{m}{A model object}
}
\value{
Four Plots displaying the Goodness of Fit
}
\description{
Plots the Goodness of fit for the output of a model of clas `ergm`
}
\details{
This function processes the output of unipartite Exponential Random Graphs Models only. 

It does not process bipartite ERGMs model objects nor other types of Exponential random graph models object.
}
\examples{
\dontrun{
flo <-SNA4DSData::florentine
fflom <- flo$flomarriage
flom <- intergraph::asNetwork(fflom)
m <- ergm::ergm(flom ~ edges + nodecov("Wealth"))

GOF.plot(m)
}
}
