% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{ia_enron}
\alias{ia_enron}
\title{ia_enron network data}
\format{
143 vertices and 622 edges. Directed, Weighted.
}
\usage{
ia_enron
}
\description{
ia_enron network
}
\details{
ia_enron network, in \code{igraph}-format. This network is used in the course
text book.

A subset of the enron network, used in the book.
}
\keyword{datasets}
