% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_index_code.R
\name{make_index_code}
\alias{make_index_code}
\title{Create code for index.Rmd}
\usage{
make_index_code(
  rmd_file,
  lecture_description = "Week 02 -- Lab: Intro to R in SNA"
)
}
\arguments{
\item{rmd_file}{the full path to the Rmd file that needs to be parsed (typically a path on your own computer)}

\item{lecture_description}{the title for the slide deck}
}
\value{
code to be copied as is into index.rmd
}
\description{
Creates the code to go into index.Rmd
}
\details{
This function searches through the slide Rmd file and extracts the 
tags ("names") and accompanying descriptions. 
Then, if creates the full code that only needs to be copied and pasted as is 
into index.rmd to have fully up-to-date links. 
Rendering index.Rmd yields a fully functional index.html file.

Of course, it depends on providing this function with the correct info.

NOTE: using the 'name' option in a xaringan Rmd does not work well when done 
on a slide that also has \code{layout: true}. 
In this case, leave the slide with \code{layout: true} empty in terms of 
content and use \code{name} on the very next slide where you do have content.
Otherwise, many (all?) browser can not find the shortcut and will open the 
frontpage of the slidedeck, rather than the slide \code{name} refers to.

As a useful convention for the title of the deck in index.html (this is 
given in \code{lecture_description}) is:

Week 02 -- Lab: Intro to R in SNA
}
\keyword{internal}
