% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{northwestern}
\alias{northwestern}
\title{Northwestern University Facebook network}
\format{
A data frame with 10567 vertices and 488336 edges. Undirected, unweighted.
}
\source{
\url{http://networkrepository.com/socfb-Northwestern25.php}
}
\usage{
data(northwestern)
}
\description{
Northwestern University Facebook network
}
\details{
Northwestern University Facebook network, in \code{igraph}-format. This network
is used in the course text book.

 A social friendship network extracted from Facebook consisting of people
 (nodes) with edges representing friendship ties.
}
\references{
\insertRef{Traud:2011fs}{SNA4DS}

\insertRef{traud2012social}{SNA4DS}
}
\keyword{datasets}
