% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_SNA4DS_vignettes_from_internet.R
\name{open_SNA4DS_vignettes_from_internet}
\alias{open_SNA4DS_vignettes_from_internet}
\title{Open a SNA4DS vignette from the internet}
\usage{
open_SNA4DS_vignettes_from_internet(graphics = TRUE)
}
\arguments{
\item{graphics}{logical, should the list of options be shown as a clickable
graphical menu?}
}
\value{
nothing
}
\description{
Open a SNA4DS vignette using an active internet connection
}
\details{
Shows a list of currently available online vignettes for the #' \code{SNA4DS} package.
The user can pick the preferred vignette by entering the number that corresponds
to the preferred vignette. The vignette will then open in the user's default
web browser.

\emph{NOTE: this requires an active internet connection.}

When \code{graphics} is \code{TRUE}, a graphical choice menu is shown. If that
is not preferred, or if the user's machine lacks the graphical tools needed,
setting \code{graphics} to \code{FALSE} will show the list of vignettes in the
R console.
}
\examples{
\dontrun{
open_SNA4DS_vignettes_from_internet()
}
}
