% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{routers}
\alias{routers}
\title{routers network}
\format{
A data frame with 190914 vertices and 607610 edges. Undirected, unweighted.
}
\source{
\url{http://networkrepository.com/tech-RL-caida.php}
}
\usage{
data(routers)
}
\description{
routers network
}
\details{
routers network, in \code{igraph}-format. This network
is used in the course text book. The network is known as RL-CAIDA.
}
\references{
\insertRef{rossi2013topology}{SNA4DS}

\insertRef{spring2002measuring}{SNA4DS}
}
\keyword{datasets}
