% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNA4DS-package.r
\docType{package}
\name{SNA4DS-package}
\alias{SNA4DS}
\alias{SNA4DS-package}
\title{SNA4DS: Social Network Analysis for Data Science}
\description{
This package contains the tutorials for the JADS course "Social Network Analysis for Data Science" (2020-2021). Update it frequently, to have access to the most up-to-date tutorials.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SNAnalyst/SNA4DS}
  \item Report bugs at \url{https://github.com/SNAnalyst/SNA4DS/issues}
}

}
\author{
\strong{Maintainer}: Roger Leenders \email{r.t.a.j.leenders@jads.nl}

Authors:
\itemize{
  \item Claudia Zucca \email{Claudia.Zucca@jads.nl}
}

}
\keyword{internal}
