% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNA4DS-package.r
\docType{package}
\name{SNA4DS}
\alias{SNA4DS}
\title{SNA4DS}
\description{
Package supporting the SNA4DS course
}
\details{
Functions, datasets, and tutorials, useful for the SNA4DS course.
}
\keyword{package}
