% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SNA4DS_tutorials.R
\name{SNA4DS_tutorials}
\alias{SNA4DS_tutorials}
\title{Pick a tutorial}
\usage{
SNA4DS_tutorials(graphics = TRUE)
}
\arguments{
\item{graphics}{logical, should the list of options be shown as a clickable
graphical menu?}
}
\value{
nothing
}
\description{
Pick a tutorial from the SNA4DS package
}
\details{
Shows a list of currently available tutorials in the #' \code{SNA4DS} package.
The user can pick the preferred tutorial by entering the number that corresponds
to the preferred tutorial. The tutorial will then open in the user's default
web browser.

When \code{graphics} is \code{TRUE}, a graphical choice menu is shown. If that
is not preferred, or if the user's machine lacks the graphical tools needed,
setting \code{graphics} to \code{FALSE} will show the list of tutorials in the
R console.
}
\examples{
\dontrun{
SNA4DS_tutorials()
}
}
