% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{world_air}
\alias{world_air}
\title{world air network}
\format{
A data frame with 3179 vertices and 18617 edges. Undirected, unweighted.
}
\usage{
world_air
}
\description{
world air network
}
\details{
world air network, in \code{igraph}-format. This network
is used in the course text book.

According to the book's website, the data was generated from the Feb 2, 2017
data release \url{https://github.com/jpatokal/openflights/tree/5d623a6969a1adee7961cf1c9a8a212c4a784713}.
}
\keyword{datasets}
