% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{yeast}
\alias{yeast}
\title{yeast network data}
\format{
A data frame with 1870 vertices and 4480 edges. Undirected, unweighted.
}
\source{
\url{http://networkrepository.com/bio-yeast-protein-inter.php}
}
\usage{
yeast
}
\description{
Yeast network
}
\details{
Yeast network, in \code{igraph}-format. This network is used in the course
text book.
}
\references{
\insertRef{jeong2001lethality}{SNA4DS}
}
\keyword{datasets}
